Added managing gate.

+++ app/Http/Controllers/StepController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Cell;
 use App\Enums\CellStatus;
 use App\Step;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+
 
 class StepController extends Controller
 {
@@ -36,6 +39,10 @@ class StepController extends Controller
      */
     public function store(Request $request)
     {
+        if (Gate::denies('manage')) {
+            return redirect(url('/'));
+        }
+
         $request->validate([
             'name' => 'required',
             'person' => 'required',
@@ -47,7 +54,28 @@ class StepController extends Controller
         $step = new Step($request->all());
         $step->save();
 
+        $this->changeCellStatus($step);
+
         return redirect(route('cells', $step->cell_id));
+    }
+
+    public function changeCellStatus($step) {
+        $cell = Cell::where('id', $step->cell_id)->first();
+
+        $all_steps = $cell->steps;
+
+        $completed_counter = 0;
+
+        foreach ($all_steps as $step) {
+            if ($step->status === CellStatus::COMPLETE) {
+                $completed_counter++;
+            }
+        }
+
+        if ($completed_counter === count($all_steps)) {
+            $cell->status = CellStatus::COMPLETE;
+            $cell->save();
+        }
     }
 
     /**
@@ -81,6 +109,10 @@ class StepController extends Controller
      */
     public function update(Request $request, Step $step)
     {
+        if (Gate::denies('manage')) {
+            return redirect(route('cells', $step->cell_id));
+        }
+
         $request->validate([
             'name' => 'required',
             'person' => 'required',
@@ -90,6 +122,8 @@ class StepController extends Controller
         ]);
 
         $step->update($request->all());
+
+        $this->changeCellStatus($step);
 
         return redirect(route('cells', $step->cell_id));
     }
@@ -103,6 +137,10 @@ class StepController extends Controller
      */
     public function destroy(Step $step)
     {
+        if (Gate::denies('manage')) {
+            return redirect(route('cells', $step->cell_id));
+        }
+
         $cell_id = $step->cell_id;
         $step->delete();
